requirements.txt
influxdb>=4.0.0
pandas>=0.19.2
pyepics>=3.2.6